refactor(tests): try to reduce repetition for decode tests
