+++ text-generation-inference/tests/helpers.py
@@ -1,0 +1,45 @@
+import os
+
+from text_generation_server.pb.generate_pb2 import (
+    NextTokenChooserParameters,
+    Request,
+    StoppingCriteriaParameters,
+)
+
+from optimum.tpu.model import fetch_model
+
+
+def prepare_model(model_id, sequence_length):
+    # Add variables to environment so they can be used in AutoModelForCausalLM
+    os.environ["HF_SEQUENCE_LENGTH"] = str(sequence_length)
+    path = fetch_model(model_id)
+    return path
+
+
+def create_request(
+    id: int,
+    inputs: str,
+    max_new_tokens=20,
+    do_sample: bool = False,
+    top_k: int = 50,
+    top_p: float = 0.9,
+    temperature: float = 1.0,
+    seed: int = 0,
+    repetition_penalty: float = 1.0,
+):
+    # For these tests we can safely set typical_p to 1.0 (default)
+    typical_p = 1.0
+    if not do_sample:
+        # Drop top_p parameter to avoid warnings
+        top_p = 1.0
+    parameters = NextTokenChooserParameters(
+        temperature=temperature,
+        top_k=top_k,
+        top_p=top_p,
+        do_sample=do_sample,
+        seed=seed,
+        repetition_penalty=repetition_penalty,
+        typical_p=typical_p,
+    )
+    stopping_parameters = StoppingCriteriaParameters(max_new_tokens=max_new_tokens)
+    return Request(id=id, inputs=inputs, parameters=parameters, stopping_parameters=stopping_parameters)
